requirements.txt
Pillow>=10.0
fpdf2>=2.7
numpy>=1.24
imagehash>=4.3
